first/student/pyproject.toml
[tool.poetry]
name = "first-teacher"
version = "0.1.0"
description = ""
authors = ["Juan Lee <me@juany.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
